+++ quill/quill/include/quill/PatternFormatter.h
@@ -168,14 +168,17 @@ public:
    * Constructor for a PatterFormater with a custom format
    * @tparam TConstantString
    * @param format_pattern format_pattern a format string. Must be passed using the macro QUILL_STRING("format string");
+   * @param date_format The for format of the date. Same as strftime() format.
+   * @param timezone The timezone of the timestamp
    */
   template <typename TConstantString>
-  PatternFormatter(TConstantString format_pattern, std::string date_format, TimestampPrecision timestamp_precision)
+  PatternFormatter(TConstantString format_pattern, std::string date_format, TimestampPrecision timestamp_precision, Timezone timezone)
   {
     _set_pattern(format_pattern);
 
     _date_format = std::move(date_format);
     _timestamp_precision = timestamp_precision;
+    _timezone_type = timezone;
   }
 
   /**
@@ -387,8 +390,8 @@ private:
   mutable fmt::wmemory_buffer _w_memory_buffer;
 #endif
 
-  std::string _date_format{"%H:%M:%S"};       /** Timestamp format **/
-  Timezone _timezone_type{Timezone::GmtTime}; /** Timezone, GMT time by default **/
+  std::string _date_format{"%H:%M:%S"};         /** Timestamp format **/
+  Timezone _timezone_type{Timezone::LocalTime}; /** Timezone, GMT time by default **/
   TimestampPrecision _timestamp_precision{TimestampPrecision::NanoSeconds}; /** timestamp precision */
 };